pyproject.toml
[project]
name = "aise"
version = "0.0.1"
description = "AI4SE coding agent (prototype): generate and maintain executable CodeWiki aligned with Git."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "jsonschema>=4.21.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.1",
  "keyring>=25.0.0",
  "cryptography>=42.0.0",
]

[project.scripts]
aise = "aise.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"aise.resources.schemas" = ["*.json"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
